timesheets_v3: Add new Timesheets Excel templates (v2.0.3 and v3.0.3)
v3.x.x moves the Raw End Time column to after the Comments and
Reference Number.  It's easier to enter time as you go.

v2.x.x retains the Raw End Time column before the Comments column.
It's easier for entering time retrospectively.

vx.0.3 applies the same changes to both templates, with changes to the
formulae allowing the first timesheet row to have the same formulae as
subsequent rows.  This allows the first row to be deleted, if needed,
without breaking the spreadsheet.